Update to show IoC, Moq and configurable options
Extended the app to use a DryIoC container and DI for some of the classes.
The unit tests have been update to use IoC instead of creating objects manually per test
One set of unit tests continues to use a stub via IoC, while another now shows the use of Moq to generate simple mocks that are set up to return test data on specific conditions.
The game it self allow the end user to configure the size of the board using command line arguments as well as drawing a simple map of their journey at the end of each game. At the end of the game they can now choose to play again or exit.